requirements.txt
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.0.1
numpy<2.0.0
torch==2.0.1
torchvision==0.15.2
